+++ systemimage/service.py
@@ -40,7 +40,7 @@ from systemimage.main import DEFAULT_CONFIG_FILE
 # the systemimage-dev binary package is installed in Ubuntu.
 try:
     from systemimage.testing.dbus import instrument, get_service
-except ImportError:
+except ImportError: # pragma: no cover
     instrument = None
     get_service = None
 
@@ -51,6 +51,8 @@ __version__ = resource_bytes(
 
 def main():
     global config
+    # If enabled, start code coverage collection as early as possible.
+    # Parse arguments.
     parser = argparse.ArgumentParser(
         prog='system-image-dbus',
         description='Ubuntu System Image Upgrader DBus service')
@@ -66,7 +68,7 @@ def main():
                         default=0, action='count',
                         help='Increase verbosity')
     # Hidden argument for special setup required by test environment.
-    if instrument is not None:
+    if instrument is not None: # pragma: no branch
         parser.add_argument('--testing',
                             default=False, action='store',
                             help=argparse.SUPPRESS)
@@ -76,7 +78,7 @@ def main():
         config.load(args.config)
     except FileNotFoundError as error:
         parser.error('\nConfiguration file not found: {}'.format(error))
-        assert 'parser.error() does not return'
+        assert 'parser.error() does not return' # pragma: no cover
     # Load the optional channel.ini file, which must live next to the
     # configuration file.  It's okay if this file does not exist.
     channel_ini = os.path.join(os.path.dirname(args.config), 'channel.ini')
@@ -101,7 +103,7 @@ def main():
     if code == dbus.bus.REQUEST_NAME_REPLY_EXISTS:
         # Another instance already owns this name.  Exit.
         log.error('Cannot get exclusive ownership of bus name.')
-        sys.exit(2)
+        return 2
 
     log.info('SystemImage dbus main loop starting [{}/{}]',
              config.channel, config.device)
@@ -118,14 +120,14 @@ def main():
             config.dbus_service = Service(system_bus, '/Service', loop)
         try:
             loop.run()
-        except KeyboardInterrupt:
+        except KeyboardInterrupt:                   # pragma: no cover
             log.info('SystemImage dbus main loop interrupted')
-        except:
+        except:                                     # pragma: no cover
             log.exception('D-Bus loop exception')
             raise
         else:
             log.info('SystemImage dbus main loop exited')
 
 
-if __name__ == '__main__':
+if __name__ == '__main__':                        # pragma: no cover
     sys.exit(main())